Adds the option to use the extended set of GeMaps features to the modules. Changes the rounding of the features values to float32 (original one created by OpenSmile)

+++ corpus_loader.py
@@ -7,7 +7,7 @@ from sklearn import preprocessing
 
 #Assigns the analyzed_audios to the variable
 #TODO: control if the rounding of the floats is ok
-prescaled_data = (np.genfromtxt('/Users/giorgiolabate/PycharmProjects/Thesis/analyzed_audios.csv', delimiter = ',')).astype(np.float64)
+prescaled_data = (np.genfromtxt('/Users/giorgiolabate/PycharmProjects/Thesis/analyzed_audios.csv', delimiter = ',')).astype(np.float32)
 labels = np.genfromtxt('/Users/giorgiolabate/PycharmProjects/Thesis/labels.csv', delimiter = ',', dtype= str)
 #These are the possible categories of emotions
 emotions = ['Anger', 'Disgust', 'Fear', 'Joy', 'Neutrality', 'Sadness', 'Surprise']
